pyproject.toml
[project]
name = "llm-training"
version = "1.0.0"
description = "LLM勉強会 〜基礎からエージェント設計まで〜"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    # Google Generative AI SDK
    "google-genai>=1.0.0",
    # LangChain (Google Generative AI対応)
    "langchain>=0.3.14",
    "langchain-google-genai>=2.0.0",
    "langchain-core>=0.3.29",
    # 構造化出力用
    "pydantic>=2.0.0,<3.0.0",
    # 環境変数管理
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
